Move dirichlet calibration from `probly.method.dirichlet_calibration` to `probly.transformation.calibration`
