+++ src/probly/transformation/calibration/_common.py
@@ -12,7 +12,7 @@ from probly.calibrator._common import Calibrator
 from probly.predictor import BinaryLogitClassifier, BinaryProbabilisticClassifier, LogitClassifier, Predictor
 from probly.transformation.transformation import predictor_transformation
 
-type ScalingMethod = Literal["temperature", "platt", "vector", "isotonic"]
+type ScalingMethod = Literal["temperature", "platt", "vector", "isotonic", "dirichlet"]
 
 
 @dataclass(slots=True, frozen=True)
@@ -23,6 +23,8 @@ class CalibrationMethodConfig:
     vector_scale: bool = False
     use_bias: bool = False
     num_classes: int | None = None
+    reg_lambda: float = 0.0
+    reg_mu: float | None = None
 
 
 @runtime_checkable  # ty:ignore[conflicting-metaclass]
@@ -99,6 +101,56 @@ def vector_scaling[**In, Out](
     )
 
 
+@predictor_transformation(permitted_predictor_types=(LogitClassifier,), preserve_predictor_type=True)
+@CalibrationPredictor.register_factory
+def dirichlet_calibration[**In, Out](
+    base: Predictor[In, Out],
+    num_classes: int | None = None,
+    reg_lambda: float = 1e-3,
+    reg_mu: float | None = None,
+) -> CalibrationPredictor[In, Out]:
+    """Turn a multi-class logit classifier into a Dirichlet-calibrated classifier.
+
+    Based on :cite:`kullBeyondTemperatureScaling2019`.  Dirichlet calibration fits
+    a multinomial logistic regression on the log-probabilities of the base
+    classifier, ``q = softmax(W @ ln(p) + b)`` with a full ``num_classes x
+    num_classes`` weight matrix ``W`` and bias ``b``.  This generalises temperature
+    and vector scaling and recalibrates probabilities directly rather than logits.
+
+    The returned predictor still needs its parameters fitted on a held-out
+    calibration split via :func:`probly.calibrator.calibrate` (or the wrapper's
+    ``calibrate``/``fit`` methods) before it can be used for prediction.
+
+    Args:
+        base: Base logit classifier to be calibrated.
+        num_classes: Number of classes ``k`` the classifier predicts.  Must be
+            greater than one and match the class axis of the logits.
+        reg_lambda: Strength of the Off-Diagonal Regularisation, an L2 penalty on
+            the off-diagonal entries of ``W``.  ``0`` disables it, recovering the
+            unregularised full-matrix fit.
+        reg_mu: Strength of the Intercept Regularisation, an L2 penalty on the bias
+            ``b``.  Defaults to ``reg_lambda`` when ``None`` (the paper's ODIR
+            convention of tying the two strengths together).
+
+    Returns:
+        The Dirichlet calibration predictor, awaiting calibration.
+    """
+    if num_classes is None or num_classes <= 1:
+        msg = f"Dirichlet calibration expects num_classes > 1, but got {num_classes}."
+        raise ValueError(msg)
+    return calibration_generator(
+        base,
+        config=CalibrationMethodConfig(
+            method="dirichlet",
+            vector_scale=False,
+            use_bias=True,
+            num_classes=num_classes,
+            reg_lambda=reg_lambda,
+            reg_mu=reg_lambda if reg_mu is None else reg_mu,
+        ),
+    )
+
+
 @predictor_transformation(
     permitted_predictor_types=(BinaryLogitClassifier, BinaryProbabilisticClassifier), preserve_predictor_type=False
 )
@@ -117,6 +169,7 @@ __all__ = [
     "CalibrationPredictor",
     "_CalibrationPredictorBase",
     "calibration_generator",
+    "dirichlet_calibration",
     "isotonic_regression",
     "platt_scaling",
     "temperature_scaling",
